Added images taken during gameplay
There are 2 levels in this game, you need a score of 100 to get from level 1 to 2, then you need a score of 100 in level 2 to win the game. If your score reaches 0 (initial score in both levels is 30), then you lost the game. Dodge the monsters, use the secret tunnels, score and win the game. 

I hope you will enjoy playing this game! :-)